requirements.txt
Flask==3.1.0
gunicorn==23.0.0
reportlab==4.4.0
PyMuPDF==1.25.3
google-api-python-client==2.166.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
python-dotenv==1.1.0
